feat(numeric): read01 round 272 — declared precision runs to 1000
SPG capped NUMERIC's declared precision at 38, i128's width, so
numeric(50,10) — a column PG accepts — failed to parse outright. Values
wider than i128 have had somewhere to live since the arbitrary-precision
form landed; only the declared type could not say so. This closes round
271's recorded residual.

precision joins scale at u16, and the parser takes PG's range with PG's
wording for the two ends. On disk the pattern is the same as last
round's: catalog tag 10 now needs BOTH numbers to fit a byte, and tag 69
carries the wide pair, so nothing already written moves.

The cast path had its own copy of the typmod parser, reading each
component as a u8 with unwrap_or(0). numeric(1000,999) therefore did not
merely fail — it became the UNCONSTRAINED type and the cast quietly did
nothing at all. That is the third unwrap_or(0) of this shape found in
eight rounds; the pattern is a default standing in for a value the code
could not represent.

Three places assumed the old ceiling:

  * rescaling to a declared scale no i128 can hold reported "overflow
    rescaling NUMERIC" for a typmod PG accepts. It now hands the whole
    rescale to the arbitrary-precision form.
  * casting an arbitrary-precision value to a DECLARED numeric had no
    arm at all, so a 47-digit literal going into numeric(50,2) surfaced
    an internal storage type mismatch.
  * the precision check compares the mantissa against 10^precision,
    which saturates past 38 and so never fired for a wide precision —
    it would have accepted anything. It now counts integer digits when
    the limit cannot be expressed, which is PG's rule stated directly.

Also found: an unconstrained `numeric` column reported precision 0 and
scale 0 to information_schema where PG reports NULL for both. Zero is
not a precision a column can have; PG says "no constraint" by saying
nothing.

Recorded residual: a NEGATIVE declared scale (PG allows -1000..1000;
numeric(10,-2) rounds to hundreds) is still rejected. The bound message
already names the full range because that is what PG prints, but the
semantics — rounding to tens, hundreds — are a separate feature from
this round's width question.

Gate: 6460 passed / 0 failed, clippy at the 5 pre-existing warnings.

+++ crates/spg-sql/src/ast.rs
@@ -2198,7 +2198,8 @@ pub enum ColumnTypeName {
     /// `NUMERIC` / `NUMERIC(p)` / `NUMERIC(p, s)` — exact decimal.
     /// Bare `NUMERIC` and `NUMERIC(p)` both surface with `scale=0`.
     /// v7.39 (round 271) — scale widened to u16 alongside the value's.
-    Numeric(u8, u16),
+    /// v7.39 (round 272) — precision too: PG's runs to 1000.
+    Numeric(u16, u16),
     /// `DATE` — calendar day, no time-of-day component.
     Date,
     /// `TIMESTAMP` / `MySQL` `DATETIME` — instant with microsecond
